Add new method to resultsRepository, and resultsService

+++ src/Wordki.Api/Controllers/WordsController.cs
@@ -23,8 +23,10 @@ namespace Wordki.Api.Controllers
         }
 
         [HttpPost("add")]
-        public async Task<IActionResult> AddWord([FromBody] WordDTO wordDto){
-            if(wordDto == null){
+        public async Task<IActionResult> AddWord([FromBody] WordDTO wordDto)
+        {
+            if (wordDto == null)
+            {
                 throw new ApiException($"Parameter {nameof(wordDto)} cannot be null", ErrorCode.NullArgumentException);
             }
             var userId = await authorizer.AuthorizeAsync(Request);
@@ -33,23 +35,55 @@ namespace Wordki.Api.Controllers
         }
 
         [HttpPost("addAll")]
-        public async Task<IActionResult> AddAll([FromBody] IEnumerable<WordDTO> wordsDto){
+        public async Task<IActionResult> AddAll([FromBody] IEnumerable<WordDTO> wordsDto)
+        {
+            if (wordsDto == null)
+            {
+                throw new ApiException($"Parameter {nameof(wordsDto)} cannot be null", ErrorCode.NullArgumentException);
+            }
+            if (wordsDto.Count() == 0){
+                throw new ApiException($"Parameter {nameof(wordsDto)} cannot be empty", ErrorCode.NullArgumentException);
+            }
             var userId = await authorizer.AuthorizeAsync(Request);
             wordsDto = await wordService.AddAllAsync(wordsDto, userId);
             return Json(wordsDto);
         }
 
-        [HttpDelete("remove")]
-        public async Task<IActionResult> Remove([FromBody] long wordId){
+        [HttpPut("remove")]
+        public async Task<IActionResult> Remove([FromBody] long wordId)
+        {
+            if(wordId == 0){
+                throw new ApiException($"Parameter {nameof(wordId)} cannot be equal 0", ErrorCode.NullArgumentException);
+            }
             var userId = await authorizer.AuthorizeAsync(Request);
             await wordService.RemoveAsync(wordId);
             return Ok();
         }
 
         [HttpPut("update")]
-        public async Task<IActionResult> Update([FromBody] WordDTO wordDto){
-            var userId = await authorizer.AuthorizeAsync(Request);
+        public async Task<IActionResult> Update([FromBody] WordDTO wordDto)
+        {
+            if (wordDto == null)
+            {
+                throw new ApiException($"Parameter {nameof(wordDto)} cannot be null", ErrorCode.NullArgumentException);
+            }
+            await authorizer.AuthorizeAsync(Request);
             await wordService.UpdateAsync(wordDto);
+            return Ok();
+        }
+
+        [HttpPut("updateAll")]
+        public async Task<IActionResult> UpdateAll([FromBody] IEnumerable<WordDTO> wordsDto)
+        {
+            if (wordsDto == null)
+            {
+                throw new ApiException($"Parameter {nameof(wordsDto)} cannot be null", ErrorCode.NullArgumentException);
+            }
+            if (wordsDto.Count() == 0){
+                throw new ApiException($"Parameter {nameof(wordsDto)} cannot be empty", ErrorCode.NullArgumentException);
+            }
+            await authorizer.AuthorizeAsync(Request);
+            await wordService.UpdateAllAsync(wordsDto);
             return Ok();
         }
     }
